Major documentation update
Documentation for running a single PECANS instance completely updated.
Made a few changes to option names in the emissions code to made things clearer.
Modified the main run script to allow the configuration to actually set the output directory.
Updated the plotting script.
Switched the doc theme to the ReadTheDocs one because some of the function docstrings
weren't rendering well in the Alabaster theme.
Added the ensemble API docstrings.